refactor: refactor Card to be more inline with other packages
