+++ packages/card/card-loading-proptypes.js
@@ -1,0 +1,11 @@
+import PropTypes from "prop-types";
+
+export const loadingPropTypes = {
+  aspectRatio: PropTypes.number,
+  showImage: PropTypes.bool
+};
+
+export const loadingDefaultProps = {
+  aspectRatio: 3 / 2,
+  showImage: false
+};
